Total refactor.
