+++ .gitignore
@@ -286,7 +286,10 @@ __pycache__/
 *.btm.cs
 *.odx.cs
 *.xsd.cs
+
 /aws_tester
-/Frends.Community.AWS.Download/*.exe
-/Frends.Community.AWS.Helpers/NuGet.exe
-/Frends.Community.AWS.Helpers
+/aws_tester/Properties/AssemblyInfo.cs
+/aws_tester/Program.cs
+/aws_tester/packages.config
+/aws_tester/aws_tester.csproj
+/aws_tester/App.config
