+++ Frends.Community.AWS.Tests/packages.config
@@ -1,5 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="MSTest.TestAdapter" version="1.1.11" targetFramework="net452" />
+  <package id="MSTest.TestFramework" version="1.1.11" targetFramework="net452" />
   <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net452" />
   <package id="NUnit" version="3.9.0" targetFramework="net452" />
+  <package id="NUnit3TestAdapter" version="3.9.0" targetFramework="net452" />
 </packages>
