--- Frends.Community.AWS.Upload/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="AWSSDK.Core" version="3.3.20" targetFramework="net452" />
-  <package id="AWSSDK.S3" version="3.3.15" targetFramework="net452" />
-  <package id="Frends.Tasks.Attributes" version="1.2.1" targetFramework="net452" />
-</packages>
