+++ Frends.Community.AWS.sln
@@ -3,13 +3,22 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26228.9
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.AWS.Upload", "Frends.Community.AWS.Upload\Frends.Community.AWS.Upload.csproj", "{F1E9A4F4-9A5A-41ED-BF69-C50D6F443673}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.AWS.Tests", "Frends.Community.AWS.Tests\Frends.Community.AWS.Tests.csproj", "{56768113-F0F5-4F4B-BD62-A37EC863FDAE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.AWS.Download", "Frends.Community.AWS.Download\Frends.Community.AWS.Download.csproj", "{DD893081-F509-4DC1-A720-9704791387C5}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution items", "Solution items", "{B8D1CCE0-861D-4293-9104-60D56ED2907F}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		Frends.Community.AWS\FrendsTaskMetadata.json = Frends.Community.AWS\FrendsTaskMetadata.json
+		LICENSE = LICENSE
+		README.md = README.md
+	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.AWS.Listing", "Frends.Community.AWS.Listing\Frends.Community.AWS.Listing.csproj", "{858F258A-0B61-4EAE-BA59-B6E80F03AE16}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.AWS", "Frends.Community.AWS\Frends.Community.AWS.csproj", "{FCACA2AC-05A9-40E5-9A20-2B55E10075A2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Frends.Community.AWS.Tests", "Frends.Community.AWS.Tests\Frends.Community.AWS.Tests.csproj", "{64B0CEE2-1121-482E-A227-AC932D827ABD}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "nuspec", "nuspec", "{F25A1CBD-321B-445D-B6EA-631118B84CA1}"
+	ProjectSection(SolutionItems) = preProject
+		nuspec\Frends.Community.AWS.nuspec = nuspec\Frends.Community.AWS.nuspec
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,24 +26,19 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F1E9A4F4-9A5A-41ED-BF69-C50D6F443673}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F1E9A4F4-9A5A-41ED-BF69-C50D6F443673}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F1E9A4F4-9A5A-41ED-BF69-C50D6F443673}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F1E9A4F4-9A5A-41ED-BF69-C50D6F443673}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DD893081-F509-4DC1-A720-9704791387C5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DD893081-F509-4DC1-A720-9704791387C5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DD893081-F509-4DC1-A720-9704791387C5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DD893081-F509-4DC1-A720-9704791387C5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{858F258A-0B61-4EAE-BA59-B6E80F03AE16}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{858F258A-0B61-4EAE-BA59-B6E80F03AE16}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{858F258A-0B61-4EAE-BA59-B6E80F03AE16}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{858F258A-0B61-4EAE-BA59-B6E80F03AE16}.Release|Any CPU.Build.0 = Release|Any CPU
-		{64B0CEE2-1121-482E-A227-AC932D827ABD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{64B0CEE2-1121-482E-A227-AC932D827ABD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{64B0CEE2-1121-482E-A227-AC932D827ABD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{64B0CEE2-1121-482E-A227-AC932D827ABD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{56768113-F0F5-4F4B-BD62-A37EC863FDAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{56768113-F0F5-4F4B-BD62-A37EC863FDAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{56768113-F0F5-4F4B-BD62-A37EC863FDAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{56768113-F0F5-4F4B-BD62-A37EC863FDAE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FCACA2AC-05A9-40E5-9A20-2B55E10075A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FCACA2AC-05A9-40E5-9A20-2B55E10075A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FCACA2AC-05A9-40E5-9A20-2B55E10075A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FCACA2AC-05A9-40E5-9A20-2B55E10075A2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{F25A1CBD-321B-445D-B6EA-631118B84CA1} = {B8D1CCE0-861D-4293-9104-60D56ED2907F}
 	EndGlobalSection
 EndGlobal
